Add starting script
I have made a function that prints a bubble with some text
inside it. It successfully takes the bash escape code for
current directory, I haven't tested it with others. I've also
started making variables for colors, and I have set up the prompt
to be a nice-to-maintain multi-line string.