update(git_prompt): improve speed, other refactors
Improve speed by storing the git status in a variable too, and only
re-running the git status command if that variable is empty. It should
be empty on the first time it needs to run when you're making a new
prompt, so if it were to be called subsequent times in the same prompt
generation, it would be able to just use the variable instead.

Update the usage message to only have one echo, and be more clear with
the usage of the script. Since it does something with no arguments now,
I needed to change up the usage message and remove the exit condition
for no arguments.

Change variable names to be prefixed GIT_PROMPT. Also change control
variable GIT_PROMPT to be named ENABLE_GIT_PROMPT instead.