update(battery_percent): split script into two
One script just for the battery percent, one that takes a battery
percent and prints a color. Add a nice error message for the color
script, and quit at proper times if you don't need to continue. Also
creates variables that can be defined in the theme, and they won't be
overwritten. Neither script fails if the file for the information they
need doesn't exist.