fix(theme): integer comparison bug in default
Change comparison to strings, because int comparison with an unset
variable causes an error.

Also change $ to \$ in the string, because why not escape it if you plan
on it just being a dollar sign. If someone's reading this that knows for
sure that it's better to leave it with just $, then let me know.