feat(ept): edit inputrc in preview window
Apply check to see if the $EDITOR variable is vim or nvim, and opens the
prompt's inputrc in the preview window if it is.

Add switch case to open the inputrc in a split window in any different
editor that supports it. If anyone has more editors that they would like
me to add; you can open a pull request, or you could make a feature
request in the "issues", or you could just email me. I'll add anything
that supports it! :)