fix(check_kernel_behind): exit correctly
Exit 1 instead of 0 if we can't find pacman, to indicate that the kernel
is not behind (more like we can't check because I don't have a reason to
make this more portable yet). Also looks for pacman with /usr/bin/pacman
instead of which pacman, to avoid other programs appearing under which.